Fix DNS record editing
Fixes #39

When a record is edited the recordTypeSpinner should be moved to the
matching record type.  In order to refer to the values by position index
we need to ensure the order of the values.  Since a HashMap does not
ensure the order of the values, the position of the 'A' record for
instance is not predictable.  By converting to a LinkedHashMap the order
is ensured and we can set the selection by position.